[Behat] Extract steps from ProductContext to ProductAttributeContext
