+++ src/Sylius/Behat/Context/Ui/Shop/ProductContext.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
 namespace Sylius\Behat\Context\Ui\Shop;
 
 use Behat\Behat\Context\Context;
-use Behat\Mink\Element\NodeElement;
 use Sylius\Behat\Element\Product\IndexPage\VerticalMenuElementInterface;
 use Sylius\Behat\Page\ErrorPageInterface;
 use Sylius\Behat\Page\Shop\Product\IndexPageInterface;
@@ -184,88 +183,6 @@ final class ProductContext implements Context
     public function iShouldBeOnProductDetailedPage(ProductInterface $product): void
     {
         Assert::true($this->showPage->isOpen(['slug' => $product->getSlug()]));
-    }
-
-    /**
-     * @Then I should (also) see the product attribute :attributeName with value :expectedAttribute
-     */
-    public function iShouldSeeTheProductAttributeWithValue($attributeName, $expectedAttribute): void
-    {
-        Assert::same($this->showPage->getAttributeByName($attributeName), $expectedAttribute);
-    }
-
-    /**
-     * @Then /^I should(?:| also) see the product attribute "([^"]+)" with (positive|negative) value$/
-     */
-    public function iShouldSeeTheProductAttributeWithBoolean($attributeName, $expectedAttribute): void
-    {
-        Assert::same($this->showPage->getAttributeByName($attributeName), 'positive' === $expectedAttribute ? 'Yes' : 'No');
-    }
-
-    /**
-     * @Then I should (also) see the product attribute :attributeName with value :expectedAttribute on the list
-     */
-    public function iShouldSeeTheProductAttributeWithValueOnTheList($attributeName, $expectedAttribute): void
-    {
-        Assert::inArray($expectedAttribute, $this->showPage->getAttributeListByName($attributeName));
-    }
-
-    /**
-     * @Then I should not see the product attribute :attributeName
-     */
-    public function iShouldNotSeeTheProductAttribute(string $attributeName): void
-    {
-        $this->showPage->getAttributeByName($attributeName);
-    }
-
-    /**
-     * @Then I should (also) see the product attribute :attributeName with date :expectedAttribute
-     */
-    public function iShouldSeeTheProductAttributeWithDate($attributeName, $expectedAttribute): void
-    {
-        Assert::eq(
-            new \DateTime($this->showPage->getAttributeByName($attributeName)),
-            new \DateTime($expectedAttribute)
-        );
-    }
-
-    /**
-     * @Then /^I should(?:| also) see the product attribute "([^"]+)" with value ([^"]+)%$/
-     */
-    public function iShouldSeeTheProductAttributeWithPercentage(string $attributeName, int $expectedAttribute): void
-    {
-        Assert::eq(
-            $this->showPage->getAttributeByName($attributeName),
-            sprintf('%d %%', $expectedAttribute)
-        );
-    }
-
-    /**
-     * @Then I should see :count attributes
-     */
-    public function iShouldSeeAttributes($count): void
-    {
-        Assert::same(count($this->getProductAttributes()), (int) $count);
-    }
-
-    /**
-     * @Then the first attribute should be :name
-     */
-    public function theFirstAttributeShouldBe($name): void
-    {
-        $attributes = $this->getProductAttributes();
-
-        Assert::same(reset($attributes)->getText(), $name);
-    }
-
-    /**
-     * @Then the last attribute should be :name
-     */
-    public function theLastAttributeShouldBe($name): void
-    {
-        $attributes = $this->getProductAttributes();
-
-        Assert::same(end($attributes)->getText(), $name);
     }
 
     /**
@@ -997,18 +914,5 @@ final class ProductContext implements Context
                 $productAssociationName
             )
         );
-    }
-
-    /**
-     * @return NodeElement[]
-     *
-     * @throws \InvalidArgumentException
-     */
-    private function getProductAttributes(): array
-    {
-        $attributes = $this->showPage->getAttributes();
-        Assert::notNull($attributes, 'The product has no attributes.');
-
-        return $attributes;
     }
 }
