+++ src/Sylius/Behat/Resources/config/suites/api/product/viewing_products.yml
@@ -29,6 +29,7 @@ default:
 
                 - sylius.behat.context.api.shop.channel
                 - sylius.behat.context.api.shop.product
+                - sylius.behat.context.api.shop.product_attribute
                 - sylius.behat.context.api.shop.product_variant
 
             filters:
