+++ src/Sylius/Behat/Resources/config/suites/ui/product/viewing_products.yml
@@ -44,6 +44,7 @@ default:
                 - sylius.behat.context.ui.channel
                 - sylius.behat.context.ui.shop.locale
                 - sylius.behat.context.ui.shop.product
+                - sylius.behat.context.ui.shop.product_attribute
                 - sylius.behat.context.ui.shop.browsing_product
 
             filters:
